Move generate meeting/decision-numbers to period.
