+++ opengever/meeting/model/period.py
@@ -81,13 +81,5 @@ class Period(Base, SQLFormSupport):
     def can_execute_transition(self, name):
         return self.workflow.can_execute_transition(self, name)
 
-    def get_next_decision_sequence_number(self):
-        self.decision_sequence_number += 1
-        return self.decision_sequence_number
-
-    def get_next_meeting_sequence_number(self):
-        self.meeting_sequence_number += 1
-        return self.meeting_sequence_number
-
     def get_toc_template(self):
         return self.committee.get_toc_template()
