+++ opengever/meeting/period.py
@@ -32,3 +32,11 @@ class Period(Container):
             raise MultiplePeriodsFound()
 
         return brains[0].getObject()
+
+    def get_next_decision_sequence_number(self):
+        self.decision_sequence_number += 1
+        return self.decision_sequence_number
+
+    def get_next_meeting_sequence_number(self):
+        self.meeting_sequence_number += 1
+        return self.meeting_sequence_number
